Significantly increased the resolution of the Oscilloscope screen (at the expense of performance)
OpenGL buffer is increase to the use the full read data, rather than a compression to the pixel size of the screen. This came with significant performance degration which was countered by loading only the inscreen datapoints into the openGL engine.
(performance is now 70-100ms for 10s data to 22ms for 100ms data to <16ms for 10ms data and less). Further performance increases can be obtained by disallowing scale changes during drawing (allows to move the transformation calculations to be static during drawing).